loan and collateral row refactor
